requirements.txt
fastapi==0.104.1
paddleocr==3.3.1
pillow==10.1.0
uvicorn==0.24.0
structlog==23.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
